Moved UDPConfigHandler to its own classfile
