+++ udpSource/src/main/java/io/vantiq/extsrc/udp/ConfigurableUDPSource.java
@@ -21,7 +21,6 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.*;
-import java.util.regex.Pattern;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.slf4j.Logger;
@@ -136,228 +135,7 @@ public class ConfigurableUDPSource {
      * Creates publish and notification handlers for any messages relating to a source based on the configuration
      * document
      */
-    static Handler<ExtensionServiceMessage> UDPConfig = new Handler<ExtensionServiceMessage>() {
-        @Override
-        public void handleMessage(ExtensionServiceMessage message) {
-            String sourceName = message.getSourceName();
-            Map srcConfig = (Map) ((Map)message.getObject()).get("config");
-            if (!(srcConfig.get("extSrcConfig") instanceof Map)) {
-                log.error("Unable to obtain server configuration for '" + sourceName + "'. Source '" +
-                        sourceName  + "' is terminating.");
-                clients.get(sourceName).close();
-                return;
-            }
-            Map config = (Map) srcConfig.get("extSrcConfig");
-            log.trace("Creating handlers for '" + sourceName + "'");
-
-            // Acquire the general settings Map and the listening port and address
-            Map general = null;
-            if (config.get("general") instanceof Map) {
-                general = (Map) config.get("general");
-            }
-            int port = getListeningPort(general, sourceName);
-            InetAddress address = getListeningAddress(general, sourceName);
-            
-            Map incoming = null;
-            Map outgoing = null;
-            if (config.get("incoming") instanceof Map) {
-                incoming = (Map) config.get("incoming");
-            }
-            if (config.get("outgoing") instanceof Map) {
-                outgoing = (Map) config.get("outgoing");
-            }
-            
-            if (!isConfiguredToSend(outgoing) && !isConfiguredToReceive(incoming)) {
-                log.error("Source '" + sourceName + "' is not configured to send or receive. Killing the connection");
-                clients.get(sourceName).close();
-                return;
-            }
-
-            // Synchronization necessary because createUDPSocket creates a list which listenOnUDPSocket will add to.
-            // If threads change between the creation of the UDP socket and the assignation of the list or while
-            // a source is being added to the list, an error is likely to occur
-            DatagramSocket socket;
-            synchronized (this) {
-                // Create the socket that the source will use for UDP messages
-                socket = createUDPSocket(port, address, sourceName);
-                if (socket == null) { // No socket could be created for the given port and address
-                    socket = listenOnUDPSocket(port, address, sourceName);
-                }
-                if (socket == null) {
-                    log.error("Failed to obtain UDP socket at address '" + address + "' and port '" + port +
-                            "' for source '" + sourceName + "'");
-                    return;
-                }
-            }
-
-            // Setup Publish handler as the configuration document requests
-            if (isConfiguredToSend(outgoing)) {
-                UDPPublishHandler handler = new UDPPublishHandler(outgoing, socket);
-                clients.get(sourceName).setPublishHandler(handler);
-                log.debug("Publish handler created for source '" + sourceName + "'");
-            }
-
-            // Setup Notification handler as the configuration document requests
-            if (isConfiguredToReceive(incoming)) {
-                UDPNotificationHandler handler = new UDPNotificationHandler(incoming, clients.get(sourceName));
-                setNotificationHandler(handler, sourceName, incoming); // This function is part of ConfigurableUDPSource
-                log.debug("Notification handler created for source '" + sourceName + "'");
-            }
-            log.info("Source '" + sourceName + "' setup and ready to go.");
-        }
-
-        private int getListeningPort(Map general,String sourceName) {
-            Object port = null;
-            if (general != null) {
-                port = general.get("listenPort");
-            }
-            if (port instanceof Integer && (int) port >= 0 && (int) port <= 65535) {
-                return (int) port;
-            }
-            else {
-                log.debug("No valid listening port specified for source '" + sourceName + "'. Using default port '"
-                        + LISTENING_PORT + "'.");
-                return LISTENING_PORT;
-            }
-        }
-
-        private InetAddress getListeningAddress(Map general, String sourceName) {
-            Object address = null;
-            if (general != null) {
-                address = general.get("listenAddress");
-            }
-
-            if (address instanceof String) {
-                try {
-                    return InetAddress.getByName((String) address);
-                }
-                catch (UnknownHostException e) {
-                    log.warn("Requested listening address '" + address + "' specified for source '" + sourceName + "'" +
-                            "could not be found. Using default address '" + LISTENING_ADDRESS + "'.");
-                    return LISTENING_ADDRESS;
-                }
-            }
-            else {
-                log.debug("No listening address specified for source '" + sourceName + "'. Using default address '" +
-                        LISTENING_ADDRESS + "'.");
-                return LISTENING_ADDRESS;
-            }
-        }
-
-        boolean isConfiguredToSend(Map outgoing) {
-            if (outgoing == null || !(outgoing.get("targetAddress") instanceof String) ||
-                    !(outgoing.get("targetPort") instanceof Integer)) {
-                return false;
-            }
-            int port = (Integer) outgoing.get("targetPort");
-            String addr = (String) outgoing.get("targetAddress");
-            if (port < 0 || port > 65535) {
-                return false;
-            }
-            try {
-                InetAddress.getByName(addr);
-                return true;
-            } catch (Exception e) {
-                return false;
-            }
-        }
-
-        boolean isConfiguredToReceive(Map incoming) {
-            return incoming != null &&
-                    (valueIsTrue(incoming, "receiveAllPorts") || valueIsTrue(incoming, "receiveAllAddresses") ||
-                    valueIsTrue(incoming, "passPureMapIn") || valueIsTrue(incoming, "passUnspecifiedIn") ||
-                    valueIsTrue(incoming,"expectCsvIn") || valueIsTrue(incoming,"expectXmlIn") ||
-                    incoming.get("passRecAddress") instanceof String || incoming.get("passRecPort") instanceof String ||
-                    incoming.get("passBytesInAs") instanceof String ||
-                    (incoming.get("receivePorts") instanceof List  
-                            && hasValidPort((List) incoming.get("receivePorts"))) ||
-                    (incoming.get("receiveAddresses") instanceof List 
-                            && hasValidAddress((List) incoming.get("receiveAddresses"))) ||
-                    (incoming.get("receiveServers") instanceof List 
-                            && hasValidServer((List) incoming.get("receiveServers"))) ||
-                    (incoming.get("transformations") instanceof List 
-                            && hasValidTransform((List) incoming.get("transformations"))) || 
-                    (incoming.get("regexParser") instanceof Map && 
-                            isValidRegexParser((Map)incoming.get("regexParser")))
-                    );
-        }
-        
-        boolean hasValidAddress(List potentialAddresses) {
-            for (Object name : potentialAddresses) {
-                if (name instanceof String) {
-                    try {
-                        InetAddress.getByName((String) name);
-                        return true;
-                    }
-                    catch (UnknownHostException e) {
-                        
-                    }
-                }
-            }
-            return false;
-        }
-        
-        boolean hasValidPort(List potentialPorts) {
-            for (Object port : potentialPorts) {
-                if (port instanceof Integer && (int) port >= 0 && (int) port <= 65535) {
-                    return true;
-                }
-            }
-            return false;
-        }
-        
-        boolean hasValidServer(List potentialServers) {
-            for (Object server : potentialServers) {
-                if (isValidServer(server)) {
-                    List<Object> s = (List) server;
-                    String addressName = (String) s.get(0);
-                    try {
-                        InetAddress.getByName(addressName);
-                        return true;
-                    }
-                    catch (UnknownHostException e) {
-                        
-                    }
-                }
-            }
-            return false;
-        }
-        
-        boolean hasValidTransform(List potentialTransformations) {
-            for (Object obj : potentialTransformations) {
-                if (obj instanceof List) {
-                    List list = (List) obj;
-                    if (list.size() == 2 && list.get(0) instanceof String && list.get(1) instanceof String) {
-                        return true;
-                    }
-                }
-            }
-            return false;
-        }
-        
-        boolean isValidRegexParser(Map potentialParser) {
-            Pattern p;
-            try {
-                p = Pattern.compile((String) potentialParser.get("pattern"));
-            }
-            catch (Throwable t) {
-                return false;
-            }
-            if (potentialParser.get("locations") instanceof List) {
-                List locs = (List)potentialParser.get("locations");
-                int strings = 0;
-                for (Object obj : locs) {
-                    if (obj instanceof String) {
-                        strings++;
-                    }
-                }
-                if (strings == p.matcher("").groupCount()) {
-                    return true;
-                }
-            }
-            return false;
-        }
-    };
+    static UDPConfigHandler UDPConfig = new UDPConfigHandler();
     
     static void clearSourceHandlers(String sourceName) {
         ExtensionWebSocketClient client = clients.get(sourceName);
@@ -626,7 +404,7 @@ public class ConfigurableUDPSource {
      * @return          {@code true} if {@code server} is a {@link List} of size 2 whose first entry is a {@link String} and
      *                  whose second entry is an integer between 0 and 65535, inclusive
      */
-    private static boolean isValidServer(Object server) {
+    static boolean isValidServer(Object server) {
         return server instanceof List && ((List) server).size() == 2 && ((List) server).get(0) instanceof String &&
                 ((List) server).get(1) instanceof Integer && (int) ((List) server).get(1) >= 0 &&
                 (int) ((List) server).get(1) <= 65535;
@@ -663,19 +441,19 @@ public class ConfigurableUDPSource {
     /**
      * An Object used to ensure only one source edits udpSocketToSources at once
      */
-    private static Object socketLock = new Object();
+    static Object socketLock = new Object();
     /**
      * The {@link ExtensionWebSocketClient} that communicates with Vantiq
      */
-    private static Map<String, ExtensionWebSocketClient> clients = new LinkedHashMap<>();
+    static Map<String, ExtensionWebSocketClient> clients = new LinkedHashMap<>();
     /**
      * The default address for a UDP server to listen on.
      */
-    private static InetAddress LISTENING_ADDRESS = null;
+    static InetAddress LISTENING_ADDRESS = null;
     /**
      * The default port for the UDP server to listen on
      */
-    private static int LISTENING_PORT = 3141;
+    static int LISTENING_PORT = 3141;
     /**
      * Max byte size receivable through UDP servers
      */
@@ -1091,7 +869,7 @@ public class ConfigurableUDPSource {
      * @return      {@code true} if {@code map.get(key)} is either a boolean and {@code true}, or not a boolean and non-null,
      *              {@code false} otherwise
      */
-    private static boolean valueIsTrue(Map map, Object key) {
+    static boolean valueIsTrue(Map map, Object key) {
         Object val = map.get(key);
         if (val instanceof Boolean) {
             return (boolean) val;
